refactor: split models/index into separate queries and mutations
